Wait retention_hours since last deploy

+++ app/jobs/cleanup_review_app_job.rb
@@ -2,6 +2,6 @@ class CleanupReviewAppJob < ActiveJob::Base
   queue_as :default
 
   def perform(reviewapp)
-    reviewapp.destroy
+    reviewapp.destroy if reviewapp.expired?
   end
 end
